bot_app/src/requirements.txt
python-telegram-bot==21.1.1
python-telegram-bot[webhooks]
google_auth==2.29.0
google-api-python-client==2.128.0
aiogoogle==5.7.0
aiohttp==3.9.5
aiofiles==23.2.1
redis==5.0.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
cryptography==42.0.7
fastapi==0.111.0
google-generativeai==0.5.4